Move get_basic_auth_header to utils
