+++ src/apify/scrapy/__init__.py
@@ -1,3 +1,3 @@
 from .pipelines import ActorDatasetPushPipeline
 from .scheduler import ApifyScheduler
-from .utils import get_running_event_loop_id, open_queue_with_custom_client, to_apify_request, to_scrapy_request
+from .utils import get_basic_auth_header, get_running_event_loop_id, open_queue_with_custom_client, to_apify_request, to_scrapy_request
